+++ packages/terra-modal-manager/src/actions.js
@@ -5,6 +5,8 @@ import {
   POP,
   MAXIMIZE,
   MINIMIZE,
+  GAIN_FOCUS,
+  LOSE_FOCUS,
 } from './actionTypes';
 
 export function open(data) {
@@ -29,4 +31,12 @@ export function maximize(data) {
 
 export function minimize(data) {
   return { type: MINIMIZE, data };
+}
+
+export function gainFocus(data) {
+  return { type: GAIN_FOCUS, data };
+}
+
+export function loseFocus(data) {
+  return { type: LOSE_FOCUS, data };
 }