+++ packages/terra-site/src/examples/modal-manager/ContentContainer.jsx
@@ -12,6 +12,8 @@ class ContentContainer extends React.Component {
     this.goBack = this.goBack.bind(this);
     this.maximize = this.maximize.bind(this);
     this.minimize = this.minimize.bind(this);
+    this.requestFocus = this.requestFocus.bind(this);
+    this.releaseFocus = this.releaseFocus.bind(this);
   }
 
   disclose(size) {
@@ -52,6 +54,14 @@ class ContentContainer extends React.Component {
     this.props.app.minimize();
   }
 
+  requestFocus() {
+    this.props.app.requestFocus();
+  }
+
+  releaseFocus() {
+    this.props.app.releaseFocus();
+  }
+
   render() {
     const { app, identifier } = this.props;
 
@@ -60,6 +70,9 @@ class ContentContainer extends React.Component {
         <h2>Content Component</h2>
         <br />
         <h4>id: {identifier}</h4>
+        <br />
+        {app && app.releaseFocus ? <h4>Modal has lost focus!</h4> : null }
+        {app && app.requestFocus ? <h4>Modal has gained focus!</h4> : null }
         <br />
         <button className="disclose" onClick={this.disclose()}>Disclose</button>
         {identifier === 'root-component' && <button className="disclose-tiny" onClick={this.disclose('tiny')}>Disclose - Tiny</button>}
@@ -72,6 +85,8 @@ class ContentContainer extends React.Component {
         {app && app.goBack ? <button className="go-back" onClick={this.goBack}>Go Back</button> : null }
         {app && app.maximize ? <button className="maximize" onClick={this.maximize}>Maximize</button> : null }
         {app && app.minimize ? <button className="minimize" onClick={this.minimize}>Minimize</button> : null }
+        {app && app.requestFocus ? <button className="requestFocus" onClick={this.requestFocus}>Request Focus</button> : null }
+        {app && app.releaseFocus ? <button className="releaseFocus" onClick={this.releaseFocus}>Release Focus</button> : null }
       </div>
     );
   }